Use tighter types for UTF-8

+++ src/writer.c
@@ -285,7 +285,7 @@ esink(const void* buf, size_t len, SerdWriter* writer)
 static size_t
 write_character(SerdWriter* const    writer,
                 const uint8_t* const utf8,
-                size_t* const        size,
+                uint8_t* const       size,
                 SerdStatus* const    st)
 {
   char           escape[11] = {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0};
@@ -360,7 +360,7 @@ write_uri(SerdWriter* writer, const char* utf8, size_t n_bytes, SerdStatus* st)
     }
 
     // Write UTF-8 character
-    size_t size = 0;
+    uint8_t size = 0;
     len += write_character(writer, (const uint8_t*)utf8 + i, &size, st);
     i += size;
     if (*st && !(writer->flags & SERD_WRITE_LAX)) {
@@ -445,7 +445,7 @@ write_lname(SerdWriter* writer, const char* utf8, const size_t n_bytes)
      sets of valid characters. */
 
   // Write first character
-  size_t    first_size = 0U;
+  uint8_t   first_size = 0U;
   const int first = (int)parse_utf8_char((const uint8_t*)utf8, &first_size);
   if (is_PN_CHARS_U(first) || first == ':' || is_digit(first)) {
     TRY(st, esink(utf8, first_size, writer));
@@ -455,7 +455,7 @@ write_lname(SerdWriter* writer, const char* utf8, const size_t n_bytes)
 
   // Write middle and last characters
   for (size_t i = first_size; i < n_bytes;) {
-    size_t    c_size = 0U;
+    uint8_t   c_size = 0U;
     const int c      = (int)parse_utf8_char((const uint8_t*)utf8 + i, &c_size);
 
     if (is_PN_CHARS(c) || c == ':' || (c == '.' && (i + 1U < n_bytes))) {
@@ -579,7 +579,7 @@ write_text(SerdWriter* writer,
 
     if (escape_len == 0) {
       // No special escape for this character, write full Unicode escape
-      size_t size = 0;
+      uint8_t size = 0;
       write_character(writer, (const uint8_t*)utf8 + i - 1, &size, &st);
       if (st && !(writer->flags & SERD_WRITE_LAX)) {
         return st;
@@ -591,7 +591,7 @@ write_text(SerdWriter* writer,
         for (; i < n_bytes && (utf8[i] & 0x80); ++i) {
         }
       } else {
-        i += size - 1;
+        i += size - 1U;
       }
     }
   }
